+++ engine/benchmark/src/androidTest/java/com/google/android/fhir/benchmark/FhirSyncWorkerBenchmark.kt
@@ -71,7 +71,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class H_FhirSyncWorkerBenchmark {
+class FhirSyncWorkerBenchmark {
 
   @get:Rule val benchmarkRule = BenchmarkRule()
