+++ workflow/benchmark/README.md
@@ -24,8 +24,7 @@ The results will be similar to this:
 1,238,212,883   ns    trace    D_FhirJsonParserBenchmark.parseLightFhirBundle
 2,785,964,288   ns    trace    E_ElmJsonLibraryLoaderBenchmark.parseImmunityCheckCqlFromFhirLibrary
   713,779,915   ns    trace    E_ElmJsonLibraryLoaderBenchmark.parseFhirHelpersCqlFromFhirLibrary
-  121,204,232   ns    trace    F_EngineDatabaseBenchmark.createAndGet
-9,833,892,387   ns    trace    G_CqlEvaluatorBenchmark.evaluatesLibrary
+9,833,892,387   ns    trace    F_CqlEvaluatorBenchmark.evaluatesLibrary
 ```
 
 Alternatively, from the command line, run the connectedCheck to run all of the tests from specified Gradle module:
